adding initial model

+++ TicketingSystem/src/main/java/com/ticketing/model/PaymentInterface.java
@@ -2,7 +2,8 @@ package com.ticketing.model;
 
 public interface PaymentInterface {
 
-    boolean pay(int amount);
-    boolean collect(int amount);
+    boolean pay(double amount);
+
+    boolean collect(double amount);
 
 }
